pyproject.toml
[project]
name = "pose-pi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "onnxruntime-gpu",
    "opencv-python",
    "numpy",
    "websockets",
    "uvicorn",
]
